Balance changes to prim floors
Changed beauty from 3 to 2 and increased work by 25% (except for the fancy tiki floor which uses gold), and amount of hay needed by +1. Also cleaned up the code inheritance (it was located in the wood floor xml for some reason but isn't actually used there, so it was moved over to the prim floor xml instead) and in-game descriptions.