Big haygrass improvement
Ranching is rather difficult right now, especially with the changes made to milk and cheese made a couple weeks ago. The amount of grazing land you might need to sustain a herd is too much.

Haygrass needed a substantial buff. It takes about 20-25 days to fully grow in the boreal right now. The proposed changes do the following:
1.) Make haygrass grow about twice as fast, and react to soil fertility more.
2.) Grants it a second purpose - creating improved, more compact grazing fields. It will not be destroyed when an animal eats it, but rather just eaten down and grow back. A muffalo would need to eat about 4 fully grown haygrass to fill up (compared to 7 normal grass). It goes dormant in the winter and comes back in the spring, so the player can slowly work on building themselves a professional grazing field.