Kindling improvement (now firewood)
"All kindling is firewood, but not all firewood is kindling". Kindling is renamed to "firewood" to better represent how it's used in game. Before, it was no better than burning wood logs, despite being a processed fuel. Firewood is here to act as a middle ground at 2x the burn time efficiency, with coal and peat at 3x.