Tested 1.0 versions with modpack, no errors
Added in the B18 Spanish & Russian into the About for AreaUnlocker.
ZiTools Minimap addon replaced BetterMiniMap extended, and requires BetterMiniMap.
DO NOT add BetterMiniMap to an existing colony, it will spam errors. Works on new colony.
Used 1.0 public Architect Icons and updated with the textures we had in B18.
All tested and working with the modpack.