Updated Dubs Bad Hygiene.
There is no changelog for what's acutally changed dubwise has scrubbed his github, but suffice to say the code itself show that there are changes.

All I've done is coppied the matt template already laid out by Sidfu so that the new version uses SK matts, and adjusted the debuffs i,e (awful bathroom) to match the values we used previously.

Dubwise did set up a new desginator catagory as HygieneMisc which I changed all of them back into just Hygiene, and temperture where appropriate on the advise of Sidfu. Reguarding that I believe other languages might come up with an error when they attempt to translate HygieneMisc which no longer exists but it didn't show up when I loaded up my game so I only assume that shows up when you switch language in game.  I'm not sure how to update language files if that is the case. but if it does become a problem it'd only be as simple as renaming a few desginator catagories for HygieneMisc to either Hygiene or Temperture where appriopriate for the languages. (Specfically Spanish as that's the only one I've updated to the newer version supplied with the updated mod) Russian-SK & Korean-SK are untouched.

Couple of things I did note one being that there is a new research for washing machines (It got it's own research) and there was changes to how dubwise uses fertilizer which no doubt we will end up not using, but I've left it in as default that comes with the mod, as i'm not familiar with the changes you and sidfu did to them.

The new fertilizer system needs to be checked by Sid or Sky.
I think Defs\ThingDefs_Buildings\BuildingsF_Irrigation.xml line 214/215 might be wrong i wasn't sure if that's changes you've done or dubwise did. so i left it as default with what comes with the mod.