Released HSK SkyAI v.1.
The main feature of the new update is the AI's response to the threats appearing before him and his group. All humanoid creatures now react to wounds, bleeding, severe pain, hunger, fatigue, inability to move, ambient temperature, and even fainting. A pawn with serious wounds as part of a group of having field doctors will request help. With minor wounds the pawn, if he has the skill, can give himself first aid on his own. Having received a wound in combat the pawn will seek to hide better, sometimes even making an attempt to escape from the line of fire, significantly reducing his attacking zeal. A seriously wounded pawn, in the absence of help from the field doctor and the impossibility / ineffectiveness of assistance by their own forces may even try to escape. In the absence of weapons in hand and the inability to pick up weapons nearby pawn may decide to leave the battlefield. In battle, pawns tend to take a comfortable position from which to fire effectively at the enemy and be covered, help allies, try to extingush fire. When retreating a pawn may decide to evacuate a wounded comrade and the whole group will now seek to leave the battlefield in one direction without scattering across the map. They may also scatter, but only under certain conditions. Badly damaged hiding places or buildings that can explode pawns are now disinclined to consider as a good position. Pawns are now able to switch to weapons, depending on the conditions in the battle. For a distant target, a weapon with a good range is suitable, at close range something sharper, and at medium range, grenades can be used also too, but as long as there are no nearby allies. Pawns have largely learned to avoid dealing damage to allies when using grenades and other heavy weapons. Moreover, the pawns' reaction speed to the threat of explosions has increased by more than 4-5 times. The pawn and its closest allies are now able to react to threats not only while in combat, but even when traveling across the map or leaving. What exactly the reaction will depend on the state: wounds, availability of weapons, and even an estimate of the chance to escape and not be killed by the type and range of the opponent's weapon. More than 30 traits affect the chance of escaping from the field of pain, either positively or negatively. So a pawn with a good combat selection of personal traits can fight to the end. Everything that was done in the previous version Siege AI 2.0 has also undergone changes and improvements, so for example the enemy snipers can now use the "kite" technique. The enemy has learned to use siege weapons almost perfectly. Grenades, rocket launchers, grenade launchers are all now used for maximum result. Pathfinding has been greatly improved, a ban on stacks in the same cell has been made. Stealing mechanics have been improved, pawns now use inventory and can carry away much more, and the selection of possible targets to steal has been expanded. Good food, resources and even a fresh animal carcass, yes anything can come in handy! Enemies are now indifferent to the player's tamed animals, which pose no threat to them and are not seen as targets. Compatible with old savesgames, only the first time you should load will be an error. It is best to persave the game after the first run. Also, already spawned raid on map on old savegame wouldn't work correctly.

Главная особенность нового обновления, это реагирование АИ на появляющейся перед ним и  его группой угрозы. Все человекоподобные существа теперь реагируют на ранения, кровотечение, сильную боль, голод, усталость, невозможность движения, температуру окружающей среды и даже предобморочное состояние. Пешка с серьезными ранениями в составе группы имеющих полевых врачей будет запрашивать помощь. При легких ранениях пешка, при наличии навыка, может оказывать себе первую помощь самостоятельно. Получив ранение в бою пешка будет стремиться укрыться лучше, иногда даже совершая попытку уйти с  линии огня, значительно снизив своё атакующее рвение. Серьезно  раненная пешка, при отсутствии помощи со стороны полевого доктора и невозможности/неэффективности оказания помощи собственными силами может и вовсе попытаться сбежать. При отсутствии оружия в руках и невозможности подобрать оружие поблизости пешка может решить покинуть поле боя. В бою, пешки стремятся занять удобную позицию, с которой можно вести эффективный огонь по врагу и быть укрытым, помогают союзникам, стараются погасить огонь. При отступлении пешка может решить эвакуировать раненого товарища и в целом вся группа будет стремиться теперь покинуть поле боя в одном направлении не разбегаясь по всей карте. Могут и разбегаться, но только при определенных условиях. Сильно поврежденные укрытия или строения которые могут взорваться пешки теперь не склонны рассматривать в качестве хорошей позиции. Пешки теперь способны переключаться на оружие, в зависимости от условий в бою. Для удаленной цели подойдет оружие с хорошей дальностью, в ближнем - что-нибудь поострее, а на средней дистанции, можно использовать гранаты, но при условии отсутствия поблизости союзников. Пешки в значительной степени научились избегать наносить урон по союзникам при использовании гранат и другого тяжелого вооружения. Более того, скорость реакции пешек на угрозу от взрывов выросла более чем в 4-5 раз. Пешка и её ближайшие союзники теперь способны реагировать на угрозы не только будучи в состоянии боя, а даже при путешествии по карте или при уходе. Какая  именно реакция будет зависеть от состояния: ранений, наличия оружия и даже оценки шанса сбежать и не быть убитым от типа и дальности оружия соперника. На шанс уйти с поля боля влияют более 30 трейтов в положительную или отрицательную сторону. Так пешка с хорошим боевым подбором личных особенностей может биться до конца. Все что было сделано в предыдущей версии Siege AI 2.0 тоже подверглось изменениям и улучшениям, так например вражеские снайперы теперь могут применять технику "кайта". Враг научился почти в совершенстве использовать осадное оружие. Гранаты, ракетные комплексы, гранатометы все теперь используется по максимуму. Значительно был улучшен поиск путей, сделан запрет на стаки в одной клетке. Улучшена механика воровства, пешки теперь используют инвентарь и могут унести значительно больше, а также расширен подбор возможных целей для кражи. Хорошая еда, ресурсы и даже свежая туша животного, да все может пригодится! Враги теперь безразличны к прирученным животным игрока, которые не представляют для них угрозы и не рассматривают в качестве целей. Совместим со старыми сохранениями, только при первой загрузке будет ошибка. Лучше всего персохранить игру после первого запуска, а также на старом сохранении не должно быть рейда врага.