Major Update
It would be impossible to list everything.

Essentially 85% of the mods have been updated for 1.4 However lots of bugs are present. Trying to understand how every single mod works is too much for one person. I'm stuck on a few key errors.

Major Errors that need fixing:
- Seeds Please
- Combat Extended Extended Loadouts
- Architect Sense

Both CE and HAR and still in Beta for 1.4. HAR has the most visual issues at the moment especially for genes.

Stuff Added:
- Most of the latest CE Biotech patches
- Added some Fire and Tox, primitive and industriall weapons and ammo.

Stuff Removed:
- Show Draftees Weapon (Now Vanilla)
- Mod Diff (Now Vanilla)

Stuff that might need removing:
- Recipe Icons, kinda integrated into vanilla.

Replacements:
- Latest RimThemes - loading screen needs adjustments for the tips and not showing the loading symbol. However everything else is the same, except the main menu is no longer hardcoded. This means other mods can alter it.

Trying out as old mod was hard to fix:
- RimTraits - Vanilla trait colors. Essentially has the same features, does need some tweaking and hsk traits that dont have a commonality stat need adding to its list. The colours are also adjustable.
- Latest Simple Sidearms - for better compatibility and easier for me. The old reason that it wasn't used is no longer really an issue (the config lag and compatibility with other mods)

Contributions:
Thanks to Obuw for providing some updated badge icons.

Im going to try and focus on fixing the main 3 issues. Most of the HAR issues will solve itself with time and ill keep testing the latest, for fixes to the issues.