Adjusted Horses/Wooly horses for Giddyup2
Giddyup2 mod by default has a minimum requirement of a bodysize of 1.3 (can be adjusted in settings) required to be mountable. 

Horses currently were set to 1.1 (unmountable per default settings) So I've increased the horses Carry size/weight/bulk/mass to match the previously wooylyhorses as was the minimum required by "hulk" pawn bodytypes to mount them. I've also increased Woolyhorse so that they're still bigger/stockier compared to regular horses and as to not be matched - by a further 1.4/20/20/5 on size/weight/bulk/mass