Change fire arrows to be crafted with sulphur
Fire arrows should be crafted with sulphur or something like that instead of fsx or gunpowder/nitrocellulose. They aren't explosive, so using explosives to make them seems odd. And historically, arrows tipped with sulphur were lit on fire before shooting. In addition, there should be the option to substitute sulphur with prometheum as that seems thematically appropriate to tip an arrow with to make it flaming. If there was a tar material or something similar, I would also suggest that, but the closest thing is from an addon.
https://discord.com/channels/272340793174392832/1258157462960472136